refactor: move the folder into features/streamer
