+++ src/features/streamer/studio/app-activity-panel.js
